Added support for ANIM_REVERSE
Allows to define animations that go in the opposite direction of the
usual flow.
Added a couple of animations from Q2 3.21 that use ANIM_REVERSE.
Make the "laser trip bomb" to use Weapon_Generic instead of the "copy &
paste" version it used.
Modified the original Weapon_Generic function to get the proper ammo
quantity the weapon uses; useful for IT_AMMO weapons.