Fix build under Ubuntu and maybe other distros.
The old problem with some distros passing --as-needed to the linker by
default and some doesn't. With --as-needed the order of object files and
libraries matters even more. See this link for details:

  https://wiki.gentoo.org/wiki/Project:Quality_Assurance/As-needed

Passing -lm and other libraries last, e.g. after all objects, it the
same we're doing in baseq2, xatrix and rogue.